# Added salvages and transitions and cleaned up the code here and there