--- a/editor/door_editor.tscn
+++ b/editor/door_editor.tscn
@@ -7,12 +7,12 @@
 [ext_resource type="PackedScene" uid="uid://h5opptcugh0i" path="res://level_elements/doors_locks/door.tscn" id="5_c2ium"]
 [ext_resource type="PackedScene" uid="uid://dioqmta8dbkbh" path="res://editor/color_choice.tscn" id="6_qvekf"]
 
-[sub_resource type="Resource" id="Resource_7nrhl"]
+[sub_resource type="Resource" id="Resource_3glrx"]
 script = ExtResource("2_jx360")
 real_part = 1
 imaginary_part = 0
 
-[sub_resource type="Resource" id="Resource_qu70u"]
+[sub_resource type="Resource" id="Resource_c3wmn"]
 script = ExtResource("3_3plxf")
 color = 1
 magnitude = 1
@@ -24,11 +24,11 @@ position = Vector2i(7, 7)
 lock_arrangement = 0
 dont_show_lock = false
 
-[sub_resource type="Resource" id="Resource_cgc1n"]
+[sub_resource type="Resource" id="Resource_igy5x"]
 script = ExtResource("4_1c01c")
-amount = SubResource("Resource_7nrhl")
+amount = SubResource("Resource_3glrx")
 outer_color = 1
-locks = Array[ExtResource("3_3plxf")]([SubResource("Resource_qu70u")])
+locks = Array[ExtResource("3_3plxf")]([SubResource("Resource_c3wmn")])
 size = Vector2i(32, 32)
 _curses = {
 0: false,
@@ -38,12 +38,13 @@ _curses = {
 }
 glitch_color = 13
 position = Vector2i(0, 0)
+sid = -1
 
 [node name="DoorEditor" type="MarginContainer"]
 offset_right = 299.0
 offset_bottom = 602.0
 script = ExtResource("1_66io0")
-door_data = SubResource("Resource_cgc1n")
+door_data = SubResource("Resource_igy5x")
 
 [node name="Panel" type="Panel" parent="."]
 visible = false
@@ -72,7 +73,7 @@ layout_mode = 2
 unique_name_in_owner = true
 layout_mode = 2
 ignore_position = true
-door_data = SubResource("Resource_cgc1n")
+door_data = SubResource("Resource_igy5x")
 ignore_collisions = true
 
 [node name="ScrollContainer" type="ScrollContainer" parent="VBoxContainer"]
@@ -97,7 +98,7 @@ text = "Color:"
 
 [node name="ColorChoice" parent="VBoxContainer/ScrollContainer/Editors/BaseDoor/Color" instance=ExtResource("6_qvekf")]
 unique_name_in_owner = true
-custom_minimum_size = Vector2(20, 450)
+custom_minimum_size = Vector2(20, 60)
 layout_mode = 2
 size_flags_horizontal = 3
 lock_size = 20