--- a/editor/door_editor.tscn
+++ b/editor/door_editor.tscn
@@ -7,12 +7,12 @@
 [ext_resource type="PackedScene" uid="uid://h5opptcugh0i" path="res://level_elements/doors_locks/door.tscn" id="5_c2ium"]
 [ext_resource type="PackedScene" uid="uid://dioqmta8dbkbh" path="res://editor/color_choice.tscn" id="6_qvekf"]
 
-[sub_resource type="Resource" id="Resource_ulo4d"]
+[sub_resource type="Resource" id="Resource_uboaf"]
 script = ExtResource("2_jx360")
 real_part = 1
 imaginary_part = 0
 
-[sub_resource type="Resource" id="Resource_0ql0t"]
+[sub_resource type="Resource" id="Resource_c5hw8"]
 script = ExtResource("3_3plxf")
 color = 1
 magnitude = 1
@@ -24,11 +24,11 @@ position = Vector2i(7, 7)
 lock_arrangement = 0
 dont_show_lock = false
 
-[sub_resource type="Resource" id="Resource_xld5w"]
+[sub_resource type="Resource" id="Resource_gvgju"]
 script = ExtResource("4_1c01c")
-amount = SubResource("Resource_ulo4d")
+amount = SubResource("Resource_uboaf")
 outer_color = 1
-locks = Array[ExtResource("3_3plxf")]([SubResource("Resource_0ql0t")])
+locks = Array[ExtResource("3_3plxf")]([SubResource("Resource_c5hw8")])
 size = Vector2i(32, 32)
 _curses = {
 0: false,
@@ -44,7 +44,7 @@ sid = -1
 offset_right = 299.0
 offset_bottom = 602.0
 script = ExtResource("1_66io0")
-door_data = SubResource("Resource_xld5w")
+door_data = SubResource("Resource_gvgju")
 
 [node name="Panel" type="Panel" parent="."]
 visible = false
@@ -73,8 +73,8 @@ layout_mode = 2
 unique_name_in_owner = true
 layout_mode = 2
 ignore_position = true
-door_data = SubResource("Resource_xld5w")
-ignore_collisions = true
+door_data = SubResource("Resource_gvgju")
+active = false
 
 [node name="ScrollContainer" type="ScrollContainer" parent="VBoxContainer"]
 layout_mode = 2