--- a/editor/door_editor.tscn
+++ b/editor/door_editor.tscn
@@ -7,12 +7,12 @@
 [ext_resource type="PackedScene" uid="uid://h5opptcugh0i" path="res://level_elements/doors_locks/door.tscn" id="5_c2ium"]
 [ext_resource type="PackedScene" uid="uid://dioqmta8dbkbh" path="res://editor/color_choice.tscn" id="6_qvekf"]
 
-[sub_resource type="Resource" id="Resource_uboaf"]
+[sub_resource type="Resource" id="Resource_74bj6"]
 script = ExtResource("2_jx360")
 real_part = 1
 imaginary_part = 0
 
-[sub_resource type="Resource" id="Resource_c5hw8"]
+[sub_resource type="Resource" id="Resource_d6how"]
 script = ExtResource("3_3plxf")
 color = 1
 magnitude = 1
@@ -24,11 +24,11 @@ position = Vector2i(7, 7)
 lock_arrangement = 0
 dont_show_lock = false
 
-[sub_resource type="Resource" id="Resource_gvgju"]
+[sub_resource type="Resource" id="Resource_fxpxb"]
 script = ExtResource("4_1c01c")
-amount = SubResource("Resource_uboaf")
+amount = SubResource("Resource_74bj6")
 outer_color = 1
-locks = Array[ExtResource("3_3plxf")]([SubResource("Resource_c5hw8")])
+locks = Array[ExtResource("3_3plxf")]([SubResource("Resource_d6how")])
 size = Vector2i(32, 32)
 _curses = {
 0: false,
@@ -44,7 +44,7 @@ sid = -1
 offset_right = 299.0
 offset_bottom = 602.0
 script = ExtResource("1_66io0")
-door_data = SubResource("Resource_gvgju")
+door_data = SubResource("Resource_fxpxb")
 
 [node name="Panel" type="Panel" parent="."]
 visible = false
@@ -73,8 +73,7 @@ layout_mode = 2
 unique_name_in_owner = true
 layout_mode = 2
 ignore_position = true
-door_data = SubResource("Resource_gvgju")
-active = false
+door_data = SubResource("Resource_fxpxb")
 
 [node name="ScrollContainer" type="ScrollContainer" parent="VBoxContainer"]
 layout_mode = 2
@@ -98,7 +97,7 @@ text = "Color:"
 
 [node name="ColorChoice" parent="VBoxContainer/ScrollContainer/Editors/BaseDoor/Color" instance=ExtResource("6_qvekf")]
 unique_name_in_owner = true
-custom_minimum_size = Vector2(20, 60)
+custom_minimum_size = Vector2(20, 450)
 layout_mode = 2
 size_flags_horizontal = 3
 lock_size = 20