--- a/level_elements/doors_locks/door.tscn
+++ b/level_elements/doors_locks/door.tscn
@@ -16,6 +16,7 @@ custom_minimum_size = Vector2(32, 32)
 offset_right = 32.0
 offset_bottom = 32.0
 script = ExtResource("1_pj0ca")
+active = null
 
 [node name="StaticBody2D" type="StaticBody2D" parent="."]
 unique_name_in_owner = true