Hi! Here are all the changes I've made regarding salvages and transitions and some other cleanup stuff.

**Some stuff to note:**

1. The UI for salvage points is currently *very simplistic* and salvage points are not actually saved at all.
2. There has been a *huge* rewrite in how objects are places and removed and all. Put simply: there are now a few generalized functions (e.g. `remove_element`) instead of tons of variants of the same kind of function like `remove_key` or `remove_door` but also there are now tons of dictionaries for names and stuff (e.g. Level.OBJECT_TYPE_TO_CONTAINER_NAME)
3. Hovering has been reworked! `mouse_exited` and `mouse_entered` are no longer used due to being too buggy. Instead, the hover is updated in the mouse movement event (is that worse for performance? I dunno)