@L4Vo5 2024-06-12
Thanks! I'll indeed be merging this, but I'll have to revert some stuff manually. If you can, next time try to make different PRs for each change, so it's easier to pick which ones to keep? Hopefully it's not too hard on your end though.
Here are some notes/comments/bugs/problems I found/ETC, as well as pointing out the stuff I fixed myself:
- The point of `@export` variables is to allow setting them in the editor, or, to save them when saving a scene/resource as a .tscn/.tres. There's no point in, for example, `@exporting` the Level variable in a salvage_point, since that's a variable that should always be set with code by whoever is controlling the salvage point. Much less in `@exporting` the door that _it itself creates_.
- Salvage flashing isn't synced up when placing them in the editor (it only ever coincidentally syncs up because loading will load them all at once). The colors should be calculated globally, like with i-view. In fact probably right next to the i-view colors (in the rendering singleton). It would also be slightly more efficient, because hsv to rgb calculations are more expensive than derefencing a global variable. (fixed)
- Salvaged doors can overlap keys and entries. And also other salvage points that don't have a door in them. But that last one could be taken as a feature?
- In the SP editor, Type repeats "Input Point" "Output Point" a bunch of times. (fixed it). It's actually something I ran into earlier in the editor's development lol. When using OptionButton and also a `@tool` script, you should .clear() it before adding the items (if you add them through code), otherwise the items get added to the ones in the saved scene.
- Cannot hover over stuff while playing :( couldn't figure out the cause of this one.

- Level data shouldn't use a bitmap. It makes level loading slower (before accounting for the "autotiling optimization" commit), and it'll have to be replaced anyway when more tile varieties are added. Even after the autotiling optimization, loading a level where over 75% of tiles are air is slower than before the PR (and faster if it's less than 75%). At the bottom I'll attach the results of the benchmark I ran.
Ok, I changed it back to dict and it seems to be faster on all cases. Also added some other very minor optimizations.
- GameplayManager's Transition can be passed by whatever knows it, either through code or in the godot editor (also, you could've accessed it with $"../../Transition" instead of get_node(NodePath(...))). But, you can keep it in the GameplayManager scene by using a CanvasLayer, so I changed it to that.
- Speaking of the Transition, it stays around until its animation finishes even if you go back to the editor.

- Clicking stuff doesn't bring up the relevant editor tab anymore (that's a bug on my own code it seems).. Ok, looks like it was a bug both on my code and on yours. On your end, set_current_tab_control wasn't working because you were passing the main Editor object instead of the element editors.
- When setting variables using a string variable, please use .get(var_name) and .set(var_name, value) instead of [var_name]. For me personally the brackets are way more confusing. Also, always keep in mind doing it this way has a performance cost compared to accessing the variables directly. Hopefully it's not much in these cases, because the generalization is very nice.
- Is there actually a memory problem with allowing SID over 99? seems like the biggest problem would be rendering a 3-digit number. Which actually looks fine? 4-digit is obviously too much for the font size though. I changed it to 999.
- Salvage editor's salvage point stops visually updating if you've clicked on a salvage in the level. This is because clicking on something changes the editor's salvage_point_data, but doesn't update the salvage_point's salvage_point_data (fixed)
- Alt+tabbing while hovering over something keeps the hover around (and moves it to the edge of the screen) until you both tab back and pass the mouse over the level viewport (this is mostly noticeable when using multiple monitors)
- Loading a level (in this case testing.tres, since saving isn't implemented for salvages) will, at first, actually place the salvaged doors in the level. You have to start and stop the level to show the actual salvage points. (These inappropiate doors cause all sorts of issues if you try to move or delete them). (fixed by revamping how levels decide whether to load salvaged doors or not, to make it similar to how it decides to load the player or not)
- Hovering over a door and then moving the mouse over a failed salvaged door (with not enough space) will keep showing the hover text of the door you started with. (since hovering doesn't work while playing anymore, currently this can only be seen by using the above bug)
- Not sure how I feel about changing all the "ignore_collisions" variables to "active". It's way less clear IMO. You're left wondering "active in what sense?" instead of going "oh, it ignores collisions". I guess I'll leave it for now tho, maybe I'll get used to it.


Tiles benchmarks I mentioned:

BEFORE PR:
Level size: (10000, 10000), Repetitions: 20
Couldn't find save data, making a new one at user://level_saves/4072774384.tres
player_spawn_position: (398, 448)
Baseline: Default level size, no tiles
Level::reset (tiles) (2 calls): 89.000 μs total, 88.000 μs self, average: 44.500 μs total, 44.000 μs self
Normal level size, tile density 10%
Level::reset (tiles) (20 calls): 8.352 ms total, 8.331 ms self, average: 417.600 μs total, 416.550 μs self
Normal level size, tile density 50%
Level::reset (tiles) (20 calls): 51.284 ms total, 51.250 ms self, average: 2.564 ms total, 2.562 ms self
Normal level size, tile density 90%
Level::reset (tiles) (20 calls): 91.177 ms total, 91.141 ms self, average: 4.559 ms total, 4.557 ms self
Big level, no tiles
Level::reset (tiles) (20 calls): 86.437 ms total, 86.402 ms self, average: 4.322 ms total, 4.320 ms self
Big level, tile density 10%
Level::reset (tiles) (20 calls): 1.694 s total, 1.694 s self, average: 84.684 ms total, 84.681 ms self
Big level, tile density 50%
Level::reset (tiles) (20 calls): 9.962 s total, 9.962 s self, average: 498.098 ms total, 498.094 ms self
Big level, tile density 90%
Level::reset (tiles) (20 calls): 17.902 s total, 17.902 s self, average: 895.101 ms total, 895.098 ms self

BEFORE AUTOTILING OPTIMIZATION (BUT WITH BITMAP):
Level size: (10000, 10000), Repetitions: 20
player_spawn_position: (398, 448)
Baseline: Default level size, no tiles
Level::reset (tiles) (2 calls): 731.000 μs total, 728.000 μs self, average: 365.500 μs total, 364.000 μs self
Normal level size, tile density 10%
Level::reset (tiles) (20 calls): 20.482 ms total, 20.443 ms self, average: 1.024 ms total, 1.022 ms self
Normal level size, tile density 50%
Level::reset (tiles) (20 calls): 79.680 ms total, 79.629 ms self, average: 3.984 ms total, 3.981 ms self
Normal level size, tile density 90%
Level::reset (tiles) (20 calls): 142.594 ms total, 142.531 ms self, average: 7.130 ms total, 7.127 ms self
Big level, no tiles
Level::reset (tiles) (20 calls): 1.709 s total, 1.709 s self, average: 85.439 ms total, 85.435 ms self
Big level, tile density 10%
Level::reset (tiles) (20 calls): 4.204 s total, 4.204 s self, average: 210.194 ms total, 210.191 ms self
Big level, tile density 50%
Level::reset (tiles) (20 calls): 16.001 s total, 16.001 s self, average: 800.045 ms total, 800.041 ms self
Big level, tile density 90%
Level::reset (tiles) (20 calls): 27.599 s total, 27.599 s self, average: 1.380 s total, 1.380 s self

AFTER AUTOTILING OPTIMIZATION:
Level size: (10000, 10000), Repetitions: 20
player_spawn_position: (398, 448)
Baseline: Default level size, no tiles
Level::reset (tiles) (2 calls): 802.000 μs total, 800.000 μs self, average: 401.000 μs total, 400.000 μs self
Normal level size, tile density 10%
Level::reset (tiles) (20 calls): 14.188 ms total, 14.165 ms self, average: 709.400 μs total, 708.250 μs self
Normal level size, tile density 50%
Level::reset (tiles) (20 calls): 31.259 ms total, 31.234 ms self, average: 1.563 ms total, 1.562 ms self
Normal level size, tile density 90%
Level::reset (tiles) (20 calls): 53.425 ms total, 53.378 ms self, average: 2.671 ms total, 2.669 ms self
Big level, no tiles
Level::reset (tiles) (20 calls): 1.612 s total, 1.612 s self, average: 80.605 ms total, 80.601 ms self
Big level, tile density 10%
Level::reset (tiles) (20 calls): 2.641 s total, 2.641 s self, average: 132.040 ms total, 132.036 ms self
Big level, tile density 50%
Level::reset (tiles) (20 calls): 6.943 s total, 6.943 s self, average: 347.135 ms total, 347.132 ms self
Big level, tile density 90%
Level::reset (tiles) (20 calls): 11.383 s total, 11.382 s self, average: 569.128 ms total, 569.125 ms self



... CHECKING 25% DENSITY WHICH SEEMS LIKE A REASONABLE AMOUNT and also just to check if that's indeed roughly the sweet spot:
BEFORE PR:
Level size: (10000, 10000), Repetitions: 20
Couldn't find save data, making a new one at user://level_saves/2101368927.tres
player_spawn_position: (398, 448)
Baseline: Default level size, no tiles
Level::reset (tiles) (2 calls): 172.000 μs total, 166.000 μs self, average: 86.000 μs total, 83.000 μs self
Normal level size, tile density 10%
Level::reset (tiles) (20 calls): 7.216 ms total, 7.195 ms self, average: 360.800 μs total, 359.750 μs self
Normal level size, tile density 25%
Level::reset (tiles) (20 calls): 25.319 ms total, 25.290 ms self, average: 1.266 ms total, 1.264 ms self
Big level, no tiles
Level::reset (tiles) (20 calls): 22.958 ms total, 22.929 ms self, average: 1.148 ms total, 1.146 ms self
Big level, tile density 10%
Level::reset (tiles) (20 calls): 1.747 s total, 1.747 s self, average: 87.344 ms total, 87.341 ms self
Big level, tile density 25%
Level::reset (tiles) (20 calls): 4.487 s total, 4.486 s self, average: 224.325 ms total, 224.322 ms self

AFTER AUTOTILING OPTIMIZATION:
Level size: (10000, 10000), Repetitions: 20
player_spawn_position: (398, 448)
Baseline: Default level size, no tiles
Level::reset (tiles) (2 calls): 650.000 μs total, 649.000 μs self, average: 325.000 μs total, 324.500 μs self
Normal level size, tile density 10%
Level::reset (tiles) (20 calls): 12.591 ms total, 12.564 ms self, average: 629.550 μs total, 628.200 μs self
Normal level size, tile density 25%
Level::reset (tiles) (20 calls): 20.562 ms total, 20.537 ms self, average: 1.028 ms total, 1.027 ms self
Big level, no tiles
Level::reset (tiles) (20 calls): 1.623 s total, 1.623 s self, average: 81.173 ms total, 81.170 ms self
Big level, tile density 10%
Level::reset (tiles) (20 calls): 2.658 s total, 2.658 s self, average: 132.885 ms total, 132.881 ms self
Big level, tile density 25%
Level::reset (tiles) (20 calls): 4.268 s total, 4.267 s self, average: 213.378 ms total, 213.374 ms self






BENCHMARK AFTER AUTOTILING OPTIMIZATION BUT MAKING TILES A DICTIONARY AGAIN:
Level size: (10000, 10000), Repetitions: 20
player_spawn_position: (398, 448)
Baseline: Default level size, no tiles
Level::reset (tiles) (2 calls): 35.000 μs total, 33.000 μs self, average: 17.500 μs total, 16.500 μs self
Normal level size, tile density 10%
Level::reset (tiles) (20 calls): 2.845 ms total, 2.823 ms self, average: 142.250 μs total, 141.150 μs self
Normal level size, tile density 25%
Level::reset (tiles) (20 calls): 10.044 ms total, 10.004 ms self, average: 502.200 μs total, 500.200 μs self
Normal level size, tile density 50%
Level::reset (tiles) (20 calls): 17.354 ms total, 17.327 ms self, average: 867.700 μs total, 866.350 μs self
Normal level size, tile density 90%
Level::reset (tiles) (20 calls): 30.839 ms total, 30.798 ms self, average: 1.542 ms total, 1.540 ms self
Big level, no tiles
Level::reset (tiles) (20 calls): 30.854 ms total, 30.826 ms self, average: 1.543 ms total, 1.541 ms self
Big level, tile density 10%
Level::reset (tiles) (20 calls): 604.632 ms total, 604.554 ms self, average: 30.232 ms total, 30.228 ms self
Big level, tile density 25%
Level::reset (tiles) (20 calls): 1.652 s total, 1.652 s self, average: 82.607 ms total, 82.604 ms self
Big level, tile density 50%
Level::reset (tiles) (20 calls): 3.864 s total, 3.864 s self, average: 193.207 ms total, 193.203 ms self
Big level, tile density 90%
Level::reset (tiles) (20 calls): 7.921 s total, 7.920 s self, average: 396.027 ms total, 396.024 ms self




... AFTER MORE POSSIBLY UNNECESSARY OPTIMIZATIONS:
Level size: (10000, 10000), Repetitions: 20
player_spawn_position: (398, 448)
Baseline: Default level size, no tiles
Level::reset (tiles) (2 calls): 30.000 μs total, 28.000 μs self, average: 15.000 μs total, 14.000 μs self
Normal level size, tile density 10%
Level::reset (tiles) (20 calls): 2.703 ms total, 2.679 ms self, average: 135.150 μs total, 133.950 μs self
Normal level size, tile density 25%
Level::reset (tiles) (20 calls): 7.065 ms total, 7.043 ms self, average: 353.250 μs total, 352.150 μs self
Normal level size, tile density 50%
Level::reset (tiles) (20 calls): 14.561 ms total, 14.533 ms self, average: 728.050 μs total, 726.650 μs self
Normal level size, tile density 90%
Level::reset (tiles) (20 calls): 25.770 ms total, 25.746 ms self, average: 1.288 ms total, 1.287 ms self
Big level, no tiles
Level::reset (tiles) (20 calls): 25.468 ms total, 25.441 ms self, average: 1.273 ms total, 1.272 ms self
Big level, tile density 10%
Level::reset (tiles) (20 calls): 495.037 ms total, 494.972 ms self, average: 24.752 ms total, 24.749 ms self
Big level, tile density 25%
Level::reset (tiles) (20 calls): 1.326 s total, 1.326 s self, average: 66.285 ms total, 66.282 ms self
Big level, tile density 50%
Level::reset (tiles) (20 calls): 2.871 s total, 2.871 s self, average: 143.559 ms total, 143.555 ms self
Big level, tile density 90%
Level::reset (tiles) (20 calls): 5.544 s total, 5.544 s self, average: 277.204 ms total, 277.200 ms self